organize service folders
