+++ src/main/scala/krs/domain/User.scala
@@ -5,11 +5,25 @@ import io.circe._
 import io.circe.generic.auto._
 import io.circe.parser._
 
+trait UserTrait {
+  id: Int
+  name: String
+  creditScore: Int
+  outstandingLoanAmount: Double
+}
+
 case class User(
-  id: Int,
-  name: String,
-  creditScore: Int,
-  outstandingLoanAmount: Double)
+  val id: Int,
+  val name: String,
+  val creditScore: Int,
+  val outstandingLoanAmmount: Double) extends UserTrait
+
+case class UserWithOffers(
+  val id: Int,
+  val name: String,
+  val creditScore: Int,
+  val outstandingLoanAmmount: Double,
+  offers: List[Offer]) extends UserTrait
 
 trait UserRepository {
   def loadUsers(): List[User]
