+++ src/main/scala/krs/service/ServiceModule.scala
@@ -1,7 +1,7 @@
 package krs.service
 
 import krs.infrastructure.InfrastructureServerModule
-import krs.api.ServerModule
+import krs.api.service.ServerModule
 import krs.domain.DomainModule
 
 trait ServiceModule extends InfrastructureServerModule with ServerModule with DomainModule
