+++ src/test/scala/krs/TestServerModule.scala
@@ -1,6 +1,6 @@
 package krs
 
-import krs.api.ServerModule
+import krs.api.service.ServerModule
 import krs.domain.DomainModule
 import krs.infrastructure.InfrastructureServerModule
